+++ src/ContaoManager/Plugin.php
@@ -12,14 +12,33 @@ use Contao\CoreBundle\ContaoCoreBundle;
 use Contao\ManagerPlugin\Bundle\BundlePluginInterface;
 use Contao\ManagerPlugin\Bundle\Config\BundleConfig;
 use Contao\ManagerPlugin\Bundle\Parser\ParserInterface;
+use Contao\ManagerPlugin\Config\ContainerBuilder;
+use Contao\ManagerPlugin\Config\ExtensionPluginInterface;
 use HeimrichHannot\LinkCheckerBundle\HeimrichHannotContaoLinkCheckerBundle;
+use HeimrichHannot\UtilsBundle\Container\ContainerUtil;
 
-class Plugin implements BundlePluginInterface
+class Plugin implements BundlePluginInterface, ExtensionPluginInterface
 {
+    /**
+     * {@inheritdoc}
+     */
     public function getBundles(ParserInterface $parser)
     {
         return [
             BundleConfig::create(HeimrichHannotContaoLinkCheckerBundle::class)->setLoadAfter([ContaoCoreBundle::class]),
         ];
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getExtensionConfig($extensionName, array $extensionConfigs, ContainerBuilder $container)
+    {
+        return ContainerUtil::mergeConfigFile(
+            'huh_encore',
+            $extensionName,
+            $extensionConfigs,
+            __DIR__.'/../Resources/config/config_encore.yml'
+        );
     }
 }