+++ CHANGELOG.md
@@ -1,10 +1,12 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
-## [2.0.0] - 2018-03-12
+## [2.0.0] - 2018-05-09
 
 ### Fixed
 - updated dependency `heimrichhannot/contao-utils-bundle` to "^2.0"
+- fixed contao 4 bundle assets paths
+- `heimrichhannot/contao-encore-bundle` support added
 
 ## [1.0.1] - 2018-03-08